devops

+++ ggcmpy/openggcm.py
@@ -1,6 +1,3 @@
-
-
-import os
 import re
 import numpy as np
 from itertools import islice
@@ -8,7 +5,7 @@ from itertools import islice
 
 def read_grid2(filename):
     # load the cell centered grid
-    with open(filename, 'r') as fin:
+    with open(filename, "r") as fin:
         nx = int(next(fin).split()[0])
         gx = list(islice(fin, 0, nx, 1))
 
@@ -18,9 +15,9 @@ def read_grid2(filename):
         nz = int(next(fin).split()[0])
         gz = list(islice(fin, 0, nz, 1))
 
-    gx = np.array(gx, dtype='f4')
-    gy = np.array(gy, dtype='f4')
-    gz = np.array(gz, dtype='f4')
+    gx = np.array(gx, dtype="f4")
+    gy = np.array(gy, dtype="f4")
+    gz = np.array(gz, dtype="f4")
 
     return {"x": gx, "y": gy, "z": gz}
 
@@ -40,7 +37,7 @@ def _as_isotime(time):
 
     ret = [None] * len(time)
     for i, t in enumerate(time):
-        t = t.strip().upper().lstrip('UT')
+        t = t.strip().upper().lstrip("UT")
         if re.match(r"^[0-9]{2}([0-9]{2}:){3,5}[0-9]{1,2}(\.[0-9]*)?$", t):
             # Handle YYYY:MM:DD:hh:mm:ss.ms -> YYYY-MM-DDThh:mm:ss.ms
             #        YYYY:MM:DD:hh:mm:s.ms  -> YYYY-MM-DDThh:mm:s.ms
@@ -48,18 +45,18 @@ def _as_isotime(time):
             #        YYYY:MM:DD:hh:mm       -> YYYY-MM-DDThh:mm
             #        YYYY:MM:DD:hh          -> YYYY-MM-DDThh
             # -- all this _tsp nonsense is to take care of s.ms; annoying
-            _tsp = t.replace('.', ':').split(':')
+            _tsp = t.replace(".", ":").split(":")
             _tsp[0] = _tsp[0].zfill(4)
             _tsp[1:6] = [_s.zfill(2) for _s in _tsp[1:6]]
             t = ":".join(_tsp[:6])
             if len(_tsp) > 6:
                 t += "." + _tsp[6]
             # --
-            ret[i] = t[:10].replace(':', '-') + 'T' + t[11:]
+            ret[i] = t[:10].replace(":", "-") + "T" + t[11:]
 
         elif re.match(r"^[0-9]{2}([0-9]{2}:){2}[0-9]{2}$", t):
             # Handle YYYY:MM:DD -> YYYY-MM-DD
-            ret[i] = t.replace(':', '-')
+            ret[i] = t.replace(":", "-")
         else:
             ret[i] = t
 
@@ -78,12 +75,12 @@ def _as_isotime(time):
 
 
 def parse_timestring(timestr):
-    prefix = 'time='
+    prefix = "time="
     timestr.strip()
     if not timestr.startswith(prefix):
         raise ValueError("Time string '{0}' is malformed".format(timestr))
-    timestr = timestr[len(prefix):].split()
+    timestr = timestr[len(prefix) :].split()
     t = float(timestr[0])
-    t = np.timedelta64(int(1000*t), "ms")
+    t = np.timedelta64(int(1000 * t), "ms")
     uttime = np.datetime64(_as_isotime(timestr[2]))
     return t, uttime
